daily_covid19_filedownload/requirements.txt
google-cloud-storage
wget
pandas